Added suggestion for EMFILE Error resolution
A [similar issue](https://github.com/stefanpenner/ember-cli/issues/2226#issuecomment-65480717) cropped up over at ember-cli, and reminded me of an issue with grunt-contrib-watch and too many open files.  Lots of people are saying this solved the for them and it seems like a better solution than increasing ulimit or altering launchctl, so I wanted to share with grunt people.